Migrate KCES wire integers to fixed-width Go types
Map KCES fields and enum values backed by C# System.Int32 to Go int32, including indexed-object metadata, catalog records, collider payloads, presets, parts assets, and system data.

Use int64 for offsets, cursors, lengths, Unity PathIDs, and internal traversal state. Remove redundant host-int validation helpers and avoid no-op int64 conversions while retaining required standard-library and reflection API boundaries.

Enforce Int32 bounds for MessagePack values, collection headers, ContentTable virtual-file sizes, and other wire narrowing points. Propagate ContentTable.AddFile errors through production callers.

Teach schema generation to reject architecture-dependent integer types in KCES roots, publish exact fixed-width metadata, and regenerate editing schemas with Int32 and Int64 bounds.

Preserve short indexed arrays, future and nil slots, raw collider unions, native JSON text, raw MessagePack payloads, and trailing wire data. Update focused regression coverage and service callers for the fixed-width API.

+++ serialization/KCES/aba/aba_robustness_test.go
@@ -496,7 +496,7 @@ func TestAbaEmptyAndDuplicateEntryNamesRoundTripByIndex(t *testing.T) {
 		if abaFile.BlockInfo.DirectoryInfos[i].Name != entry.Name {
 			t.Fatalf("directory[%d] name = %q, want %q", i, abaFile.BlockInfo.DirectoryInfos[i].Name, entry.Name)
 		}
-		data, err := abaFile.GetFileData(i)
+		data, err := abaFile.GetFileData(int64(i))
 		if err != nil {
 			t.Fatalf("GetFileData(%d): %v", i, err)
 		}
@@ -570,7 +570,7 @@ func TestAbaSampleMetadataBounds(t *testing.T) {
 		t.Skip("no .aba test files found")
 	}
 
-	var versions = make(map[uint32]int)
+	var versions = make(map[uint32]int32)
 	var maxBlocks, maxDirs int
 	var maxEntry int64
 	for _, path := range files {
